refactor: useGetColonyMembers to own hook
