+++ src/components/common/Dialogs/AwardAndSmiteDialogs/ManageReputationContainer/helpers.ts
@@ -1,10 +1,8 @@
 import Decimal from 'decimal.js';
 import { BigNumber } from 'ethers';
 
-import { useGetMembersForColonyQuery } from '~gql';
-import { Address, Colony, Member, MemberUser } from '~types';
+import { Colony } from '~types';
 import { ManageReputationMotionPayload } from '~redux/sagas/motions/manageReputationMotion';
-import { notMaybe } from '~utils/arrays';
 
 import { FormValues } from './validation';
 
@@ -29,20 +27,4 @@ export const getManageReputationDialogPayload = (
     amount: BigNumber.from(reputationChangeAmount.toString()),
     isSmitingReputation: isSmiteAction,
   } as ManageReputationMotionPayload;
-};
-
-export const useGetColonyMembers = (colonyAddress?: Address | null) => {
-  const { data } = useGetMembersForColonyQuery({
-    skip: !colonyAddress,
-    variables: {
-      input: {
-        colonyAddress: colonyAddress ?? '',
-      },
-    },
-  });
-
-  const watchers = data?.getMembersForColony?.watchers ?? [];
-  const contributors = data?.getMembersForColony?.contributors ?? [];
-  const allMembers: Member[] = [...watchers, ...contributors];
-  return allMembers.map((member) => member.user).filter<MemberUser>(notMaybe);
 };
